+++ core/DataAccess/Model.php
@@ -9,9 +9,13 @@
 namespace Piwik\DataAccess;
 
 use Exception;
+use Piwik\Archive\ArchiveInvalidator;
+use Piwik\ArchiveProcessor\ArchivingStatus;
+use Piwik\ArchiveProcessor\Parameters;
 use Piwik\ArchiveProcessor\Rules;
 use Piwik\Common;
 use Piwik\Container\StaticContainer;
+use Piwik\Date;
 use Piwik\Db;
 use Piwik\DbHelper;
 use Piwik\Period;
@@ -29,9 +33,15 @@ class Model
      */
     private $logger;
 
+    /**
+     * @var ArchivingStatus
+     */
+    private $archivingStatus;
+
     public function __construct(LoggerInterface $logger = null)
     {
         $this->logger = $logger ?: StaticContainer::get('Psr\Log\LoggerInterface');
+        $this->archivingStatus = StaticContainer::get(ArchivingStatus::class);
     }
 
     /**
@@ -56,8 +66,9 @@ class Model
                        GROUP_CONCAT(idarchive, '.', value ORDER BY ts_archived DESC) as archives
                   FROM `$archiveTable`
                  WHERE name LIKE 'done%'
+                   AND ts_archived IS NOT NULL
                    AND `value` NOT IN (" . ArchiveWriter::DONE_ERROR . ")
-              GROUP BY idsite, date1, date2, period, name";
+              GROUP BY idsite, date1, date2, period, name HAVING count(*) > 1";
 
         $archiveIds = array();
 
@@ -77,12 +88,119 @@ class Model
                     }
 
                     list($idarchive, $value) = explode('.', $pair);
-                    $archiveIds[] = $idarchive;
+                    $archiveIds[] = $idarchive; // does not matter what the value is, the latest is usable so older archives can be purged
                 }
             }
         }
 
         return $archiveIds;
+    }
+
+    public function getPlaceholderArchiveIds($archiveTable)
+    {
+        $sql = "SELECT DISTINCT idarchive FROM `$archiveTable` WHERE ts_archived IS NULL";
+        $result = Db::fetchAll($sql);
+        $result = array_column($result, 'idarchive');
+        return $result;
+    }
+
+    public function updateArchiveAsInvalidated($archiveTable, $idSites, $allPeriodsToInvalidate, Segment $segment = null, $forceInvalidateNonexistantRanges = false)
+    {
+        // select all idarchive/name pairs we want to invalidate
+        $sql = "SELECT idarchive, idsite, period, date1, date2, `name`, `value`
+                  FROM `$archiveTable`
+                 WHERE idsite IN (" . implode(',', $idSites) . ")";
+
+        if (!empty($allPeriodsToInvalidate)) {
+            $sql .= " AND (";
+
+            $isFirst = true;
+            /** @var Period $period */
+            foreach ($allPeriodsToInvalidate as $period) {
+                if ($isFirst) {
+                    $isFirst = false;
+                } else {
+                    $sql .= " OR ";
+                }
+
+                if ($period->getLabel() == 'range') { // for ranges, we delete all ranges that contain the given date(s)
+                    $sql .= "(period = " . (int)$period->getId()
+                        . " AND date2 >= '" . $period->getDateStart()->getDatetime()
+                        . "' AND date1 <= '" . $period->getDateEnd()->getDatetime() . "')";
+                } else {
+                    $sql .= "(period = " . (int)$period->getId()
+                        . " AND date1 = '" . $period->getDateStart()->getDatetime() . "'"
+                        . " AND date2 = '" . $period->getDateEnd()->getDatetime() . "')";
+                }
+            }
+            $sql .= ")";
+        }
+
+        if ($segment) {
+            $nameCondition = "name LIKE '" . Rules::getDoneFlagArchiveContainsAllPlugins($segment) . "%'";
+        } else {
+            $nameCondition = "name LIKE 'done%'";
+        }
+
+        $sql .= " AND $nameCondition";
+
+        $archivesToInvalidate = Db::fetchAll($sql);
+        $idArchives = array_column($archivesToInvalidate, 'idarchive');
+
+        // update each archive as invalidated
+        if (!empty($idArchives)) {
+            $idArchives = array_map('intval', $idArchives);
+
+            $sql = "UPDATE `$archiveTable` SET `value` = " . ArchiveWriter::DONE_INVALIDATED . " WHERE idarchive IN ("
+                . implode(',', $idArchives) . ") AND $nameCondition";
+
+            Db::query($sql);
+        }
+
+        $doneFlag = Rules::getDoneFlagArchiveContainsAllPlugins($segment ?: new Segment('', []));
+
+        // we add every archive we need to invalidate + the archives that do not already exist to archive_invalidations.
+        // except for archives that are DONE_IN_PROGRESS.
+        $archivesToCreateInvalidationRowsFor = [];
+        foreach ($archivesToInvalidate as $row) {
+            if ($row['name'] != $doneFlag) { // only look at done flags that equal the one we are explicitly adding
+                continue;
+            }
+
+            $archivesToCreateInvalidationRowsFor[$row['idsite']][$row['period']][$row['date1']][$row['date2']] = $row['idarchive'];
+        }
+
+        $now = Date::now()->getDatetime();
+
+        $dummyArchives = [];
+        foreach ($idSites as $idSite) {
+            foreach ($allPeriodsToInvalidate as $period) {
+                if ($period->getLabel() == 'range'
+                    && !$forceInvalidateNonexistantRanges
+                ) {
+                    continue; // range
+                }
+
+                $date1 = $period->getDateStart()->toString();
+                $date2 = $period->getDateEnd()->toString();
+                $idArchive = $archivesToCreateInvalidationRowsFor[$idSite][$period->getId()][$date1][$date2] ?? null;
+
+                $dummyArchives[] = [
+                    'idarchive' => $idArchive,
+                    'name' => $doneFlag,
+                    'idsite' => $idSite,
+                    'date1' => $period->getDateStart()->getDatetime(),
+                    'date2' => $period->getDateEnd()->getDatetime(),
+                    'period' => $period->getId(),
+                    'ts_invalidated' => $now,
+                ];
+            }
+        }
+
+        $fields = ['idarchive', 'name', 'idsite', 'date1', 'date2', 'period', 'ts_invalidated'];
+        Db\BatchInsert::tableInsertBatch(Common::prefixTable('archive_invalidations'), $fields, $dummyArchives);
+
+        return count($idArchives);
     }
 
     /**
@@ -93,44 +211,22 @@ class Model
      * @return \Zend_Db_Statement
      * @throws Exception
      */
-    public function updateArchiveAsInvalidated($archiveTable, $idSites, $datesByPeriodType, Segment $segment = null)
+    public function updateRangeArchiveAsInvalidated($archiveTable, $idSites, $allPeriodsToInvalidate, Segment $segment = null)
     {
-        $idSites = array_map('intval', $idSites);
-
         $bind = array();
 
         $periodConditions = array();
-        foreach ($datesByPeriodType as $periodType => $dates) {
-            $dateConditions = array();
+        if (!empty($allPeriodsToInvalidate)) {
+            foreach ($allPeriodsToInvalidate as $period) {
+                $dateConditions = array();
 
-            if ($periodType == Period\Range::PERIOD_ID) {
-                foreach ($dates as $date) {
-                    // Ranges in the DB match if their date2 is after the start of the search range and date1 is before the end
-                    // e.g. search range is 2019-01-01 to 2019-01-31
-                    // date2 >= startdate -> Ranges with date2 < 2019-01-01 (ended before 1 January) and are excluded
-                    // date1 <= endate -> Ranges with date1 > 2019-01-31 (started after 31 January) and are excluded
-                    $dateConditions[] = "(date2 >= ? AND date1 <= ?)";
-                    $bind = array_merge($bind, explode(',', $date));
-                }
-            } else {
-                foreach ($dates as $date) {
-                    $dateConditions[] = "(date1 <= ? AND ? <= date2)";
-                    $bind[] = $date;
-                    $bind[] = $date;
-                }
-            }
+                /** @var Period $period */
+                $dateConditions[] = "(date1 <= ? AND ? <= date2)";
+                $bind[] = $period->getDateStart();
+                $bind[] = $period->getDateEnd();
 
-            $dateConditionsSql = implode(" OR ", $dateConditions);
-            if (empty($periodType)
-                || $periodType == Period\Day::PERIOD_ID
-            ) {
-                // invalidate all periods if no period supplied or period is day
-                $periodConditions[] = "($dateConditionsSql)";
-            } else if ($periodType == Period\Range::PERIOD_ID) {
-                $periodConditions[] = "(period = " . Period\Range::PERIOD_ID . " AND ($dateConditionsSql))";
-            } else {
-                // for non-day periods, invalidate greater periods, but not range periods
-                $periodConditions[] = "(period >= " . (int)$periodType . " AND period < " . Period\Range::PERIOD_ID . " AND ($dateConditionsSql))";
+                $dateConditionsSql = implode(" OR ", $dateConditions);
+                $periodConditions[] = "(period = 5 AND ($dateConditionsSql))";
             }
         }
 
@@ -179,6 +275,26 @@ class Model
         }
 
         return $deletedRows;
+    }
+
+    public function getInvalidatedArchiveIdsAsOldOrOlderThan($archive)
+    {
+        $table = ArchiveTableCreator::getNumericTable(Date::factory($archive['date1']));
+        $sql = "SELECT idarchive FROM `$table` WHERE idsite = ? AND period = ? AND date1 = ? AND date2 = ? AND `name` = ? AND `value` IN ("
+            . ArchiveWriter::DONE_INVALIDATED . ") AND idarchive <= ?";
+        $bind = [
+            $archive['idsite'],
+            $archive['period'],
+            $archive['date1'],
+            $archive['date2'],
+            $archive['name'],
+            $archive['idarchive'],
+        ];
+
+        $result = Db::fetchAll($sql, $bind);
+        $result = array_column($result, 'idarchive');
+
+        return $result;
     }
 
     public function deleteArchiveIds($numericTable, $blobTable, $idsToDelete)
@@ -232,6 +348,7 @@ class Model
                                OR name = '" . ArchiveSelector::NB_VISITS_RECORD_LOOKED_UP . "'
                                OR name = '" . ArchiveSelector::NB_VISITS_CONVERTED_RECORD_LOOKED_UP . "')
                          $timeStampWhere
+                         AND ts_archived IS NOT NULL
                      ORDER BY ts_archived DESC, idarchive DESC";
         $results = Db::fetchAll($sqlQuery, $bindSQL);
 
@@ -312,7 +429,8 @@ class Model
      */
     public function getSitesWithInvalidatedArchive($numericTable)
     {
-        $rows = Db::fetchAll("SELECT DISTINCT idsite FROM `$numericTable` WHERE name LIKE 'done%' AND value = " . ArchiveWriter::DONE_INVALIDATED);
+        $rows = Db::fetchAll("SELECT DISTINCT idsite FROM `$numericTable` WHERE `name` LIKE 'done%' AND `value` IN ("
+            . ArchiveWriter::DONE_INVALIDATED . ")");
 
         $result = array();
         foreach ($rows as $row) {
@@ -430,4 +548,102 @@ class Model
         return $result;
     }
 
+    /**
+     * Marks an archive as in progress if it has not been already. This method must be thread
+     * safe.
+     */
+    public function startArchive($invalidation)
+    {
+        $table = Common::prefixTable('archive_invalidations');
+
+        // set archive value to in progress if not set already
+        $statement = Db::query("UPDATE `$table` SET `status` = ? WHERE idinvalidation = ? AND status = ?", [
+            ArchiveInvalidator::INVALIDATION_STATUS_IN_PROGRESS,
+            $invalidation['idinvalidation'],
+            ArchiveInvalidator::INVALIDATION_STATUS_QUEUED,
+        ]);
+
+        if ($statement->rowCount() > 0) { // if we updated, then we've marked the archive as started
+            return true;
+        }
+
+        // if we didn't get anything, some process either got there first, OR
+        // the archive was started previously and failed in a way that kept it's done value
+        // set to DONE_IN_PROGRESS. try to acquire the lock and if acquired, archiving isn' in process
+        // so we can claim it.
+        $lock = $this->archivingStatus->acquireArchiveInProgressLock($invalidation['idsite'], $invalidation['date1'],
+            $invalidation['date2'], $invalidation['period'], $invalidation['name']);
+        if (!$lock->isLocked()) {
+            return false; // we couldn't claim the lock, archive is in progress
+        }
+
+        return true;
+    }
+
+    /**
+     * Gets the next invalidated archive that should be archived in a table.
+     *
+     * @param string[] $tables
+     * @param int $count
+     */
+    public function getNextInvalidatedArchive($idSite, $idInvalidationsToExclude = null)
+    {
+        $table = Common::prefixTable('archive_invalidations');
+        $sql = "SELECT idinvalidation, idarchive, idsite, date1, date2, period, `name`
+                  FROM `$table`
+                 WHERE idsite = ?";
+        $bind = [
+            $idSite,
+        ];
+
+        if (!empty($idInvalidationsToExclude)) {
+            $idInvalidationsToExclude = array_map('intval', $idInvalidationsToExclude);
+            $sql .= " AND idinvalidation NOT IN (" . implode(',', $idInvalidationsToExclude) . ')';
+        }
+
+        $sql .= " ORDER BY period ASC, date1 ASC, idinvalidation ASC LIMIT 1";
+
+        return Db::fetchRow($sql, $bind);
+    }
+
+    public function deleteInvalidations($archiveInvalidations)
+    {
+        $ids = array_column($archiveInvalidations, 'idinvalidation');
+        $ids = array_map('intval', $ids);
+
+        $table = Common::prefixTable('archive_invalidations');
+        $sql = "DELETE FROM `$table` WHERE idinvalidation IN (" . implode(', ', $ids) . ")";
+
+        Db::query($sql);
+    }
+
+    /**
+     * Returns true if there is an archive that exists that can be used when aggregating an archive for $period.
+     *
+     * @param $idSite
+     * @param Period $period
+     * @return bool
+     * @throws Exception
+     */
+    public function hasChildArchivesInPeriod($idSite, Period $period)
+    {
+        $date = $period->getDateStart();
+        while ($date->isEarlier($period->getDateEnd()->addPeriod(1, 'month'))) {
+            $archiveTable = ArchiveTableCreator::getNumericTable($date);
+
+            $sql = "SELECT idarchive
+                  FROM `$archiveTable`
+                 WHERE idsite = ? AND date1 >= ? AND date2 <= ? AND period < ? AND `name` LIKE 'done%' AND `value` = " . ArchiveWriter::DONE_OK . "
+                 LIMIT 1";
+            $bind = [$idSite, $period->getDateStart()->getDatetime(), $period->getDateEnd()->getDatetime(), $period->getId()];
+
+            $result = (bool) Db::fetchOne($sql, $bind);
+            if ($result) {
+                return true;
+            }
+
+            $date = $date->addPeriod(1, 'month'); // move to next archive table
+        }
+        return false;
+    }
 }