+++ core/DataAccess/RawLogDao.php
@@ -225,7 +225,7 @@ class RawLogDao
         $sites = Db::fetchOne("SELECT 1
                 FROM " . Common::prefixTable('log_visit') . "
                 WHERE idsite = ?
-                AND visit_last_action_time > ?
+                AND visit_last_action_time >= ?
                 AND visit_last_action_time < ?
                 LIMIT 1", array($idSite, $fromDateTime, $toDateTime));
 
@@ -450,5 +450,11 @@ class RawLogDao
         }
 
         return $columns;
+    }
+
+    public function getMinimumVisitTimeForSite($idSite)
+    {
+        $sql = "SELECT MIN(visit_last_action_time) FROM " . Common::prefixTable('log_visit') . ' WHERE idsite = ?';
+        return Db::fetchOne($sql, [$idSite]);
     }
 }