+++ tests/PHPUnit/Integration/ReleaseCheckListTest.php
@@ -284,11 +284,9 @@ class ReleaseCheckListTest extends \PHPUnit\Framework\TestCase
             'DBStats'
         );
         foreach ($pluginsShouldBeDisabled as $pluginName) {
-            if (in_array($pluginName, $this->globalConfig['Plugins']['Plugins'])) {
-                throw new Exception("Plugin $pluginName is enabled by default but shouldn't.");
-            }
+            $this->assertNotContains($pluginName, $this->globalConfig['Plugins']['Plugins'],
+                "Plugin $pluginName is enabled by default but shouldn't.");
         }
-
     }
 
     /**
@@ -558,6 +556,8 @@ class ReleaseCheckListTest extends \PHPUnit\Framework\TestCase
             $icons = implode(" ", $errors);
             $this->fail("$format format failed for following icons $icons \n");
         }
+
+        $this->assertTrue(true); // pass
     }
 
     /**