+++ tests/PHPUnit/Integration/Validator/IdSiteTest.php
@@ -32,6 +32,8 @@ class IdSiteTest extends IntegrationTestCase
         $this->validate('2');
         $this->validate(1);
         $this->validate(2);
+
+        $this->assertTrue(true); // pass
     }
 
     public function test_validate_failValueDoesNotExist()